Create AppRoute component :dizzy:
Documentation: http://randycoulman.com/blog/2016/02/02/composing-routes-in-react-router/
